Fix invite password (#346)
* add InviteT to types.d

* added password to invite controller invite entry
