+++ client/src/components/Invite/InviteListing/InviteListing.tsx
@@ -1,17 +1,11 @@
-export interface InviteT {
-  inviteCode: string;
-  invitePassword: string;
-  groups: { id: string; name: string }[];
-  createdAt: string;
-  status: string;
-}
+import { InviteT } from "../../../../types";
 
 export function InviteListing({ invite }: { invite?: InviteT }) {
   return (
     <tr>
       <td className="text-center">{invite?.status || ""}</td>
       <td className="text-center">{invite?.inviteCode || ""}</td>
-      <td className="text-center">{invite?.invitePassword || ""}</td>
+      <td className="text-center">{invite?.password || ""}</td>
       <td className="text-center">
         <ul className="w-full">
           {invite?.groups
